Filtering on Spin and deleting

+++ Shared/Models/Filter.swift
@@ -9,10 +9,10 @@ import Foundation
 
 struct Filter: Identifiable, Codable {
     let id: UUID
-    var numberOfPlayers: Int
+    var numberOfPlayers: Double
     var minimumRating: Double
 
-    init(id: UUID = UUID(), numberOfPlayers: Int, minimumRating: Double) {
+    init(id: UUID = UUID(), numberOfPlayers: Double, minimumRating: Double) {
         self.id = id
         self.numberOfPlayers = numberOfPlayers
         self.minimumRating = minimumRating
